requirements.txt
streamlit
openai
langchain
google-api-python-client
google-auth
